This is me comming back to the project after a long break. I have a rough idea about what has been done during the last commit, which is mostly cleanup and implementing my own util functions for the strings. I'm making this commit so that I don't mix my new changes with the ones I made before the long break, just in case I have forgotten something and messes something up.

+++ src/util/RaceTime.cpp
@@ -1,7 +1,7 @@
 
 #include "RaceTime.h"
 
-#include "StringBuilder.h"
+#include "StringUtil.h"
 #include <string.h>
 #include <stdlib.h>
 
@@ -146,7 +146,7 @@ unsigned int RaceTime_string_to_ms(char* time_string)
  * Returns 0 on failure
  * ----------------------------------------------------------------
  */
-char* RaceTime_ms_to_string(long time)
+char* RaceTime_ms_to_string(unsigned int time)
 {
     if (time < 0)
         return 0;
