+++ src/util/RaceTime.h
@@ -2,6 +2,6 @@
 #pragma once
 
 unsigned int RaceTime_string_to_ms(char* time_string);
-char* RaceTime_ms_to_string(long time);
+char* RaceTime_ms_to_string(unsigned int time);
 char* RaceTime_percentage_to_string(float value);
